% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm_ar.R
\name{fast_AR1_chol}
\alias{fast_AR1_chol}
\title{Generate Cholesky Decomposition for an AR(1) Covariance Structure}
\usage{
fast_AR1_chol(p, rho)
}
\arguments{
\item{p}{Integer. The dimension of the covariance matrix.}

\item{rho}{Numeric. The correlation parameter for the AR(1) covariance structure.}
}
\value{
A numeric matrix of dimension 'p x p' representing the Cholesky decomposition
of the AR(1) covariance matrix.
}
\description{
This function computes the Cholesky decomposition of a covariance matrix with
an AR(1) structure. The matrix is of dimension 'p x p'.
}
\examples{
\dontrun{
  p <- 5
  rho <- 0.5
  L <- fast_AR1_chol(p, rho)
}

}
