% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm_ex.R
\name{fast_ex_chol}
\alias{fast_ex_chol}
\title{Generate Cholesky Decomposition for an Exchangeable Covariance Structure}
\usage{
fast_ex_chol(p, rho)
}
\arguments{
\item{p}{Integer. The dimension of the covariance matrix.}

\item{rho}{Numeric. The correlation parameter for the exchangeable covariance structure.}
}
\value{
A numeric matrix of dimension 'p x p' representing the Cholesky decomposition
of the exchangeable covariance matrix.
}
\description{
This function computes the Cholesky decomposition of a covariance matrix with
an exchangeable structure. The matrix is of dimension 'p x p'.
}
\examples{
\dontrun{
  p <- 5
  rho <- 0.5
  L <- fast_ex_chol(p, rho)
}

}
