% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm_tianhai.R
\name{rmvnorm_R}
\alias{rmvnorm_R}
\title{Generate Random Multivariate Normal Samples with R functions}
\usage{
rmvnorm_R(n, mu, Sigma)
}
\arguments{
\item{n}{Integer. The number of samples to generate.}

\item{mu}{Numeric vector. The mean vector for the multivariate normal distribution.}

\item{Sigma}{Numeric matrix. The covariance matrix for the multivariate normal distribution.}
}
\value{
A numeric matrix with 'n' rows and 'p' columns, where each row represents a
  random sample from the specified multivariate normal distribution.
}
\description{
This function generates random samples from a multivariate normal distribution with
a given mean vector and covariance matrix. It is more efficient compared to 
the MASS package.
}
\examples{
\dontrun{
  n <- 100
  mu <- c(0, 0)
  Sigma <- matrix(c(1, 0.5, 0.5, 1), 2, 2)
  samples <- rmvnorm_R(n, mu, Sigma)
}

}
\seealso{
\code{\link{chol}}
}
