% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm_ar.R
\name{rmvnorm_tianhai_ar1}
\alias{rmvnorm_tianhai_ar1}
\title{Generate Random Multivariate Normal Samples with AR(1) Covariance Structure}
\usage{
rmvnorm_tianhai_ar1(n, mu, p, rho)
}
\arguments{
\item{n}{Integer. The number of samples to generate.}

\item{mu}{Numeric vector. The mean vector for the multivariate normal distribution.}

\item{p}{Integer. The dimension of the multivariate normal distribution.}

\item{rho}{Numeric. The correlation parameter for the AR(1) covariance structure.}
}
\value{
A numeric matrix with 'n' rows and 'p' columns, where each row represents a
  random sample from the specified multivariate normal distribution.
}
\description{
This function generates random samples from a multivariate normal distribution with
a given mean vector and an AR(1) covariance structure.
}
\examples{
\dontrun{
  n <- 100
  mu <- c(0, 0)
  p <- 2
  rho <- 0.5
  samples <- rmvnorm_tianhai_ar1(n, mu, p, rho)
}

}
\seealso{
\code{\link{fast_AR1_chol}}, \code{\link{mat_mult_RcppArma}}
}
